Add dockerize to docker-commpose.yml
As described in issue #67, the order of services is important,
and mainflux-core must wait for emqtt to be available on port 1883.

docker-compose do not guarantees the order of starting services,
but we must use dockerize to check if one service is avilable before
we can start other service that depends on it.

Signed-off-by: Drasko DRASKOVIC <drasko.draskovic@gmail.com>