NOISSUE - Generate Invitation Key (#87)
* feat(auth): Add invitation key generation

This commit adds the functionality to generate an invitation key in the auth service. The `invitationKey` function is responsible for generating the key and issuing an access token. It sets the key type, expiration time, and subject based on the user's domain. The generated access token is returned as part of the response.

The `invitationKey` function is called in the `Issue` function when the key type is InvitationKey. This allows for the generation of invitation keys when needed.

This feature enhances the functionality of the auth service by enabling the generation of invitation keys for user invitations.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* fix(auth): fix invitation duration

The invitation duration in the auth service has been corrected to 7 days (7 * 24 * time.Hour) instead of 24 hours.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* feat(auth): configure invitation duration by env vars

Configure invitation duration by envionment variables. The default is `168h` which is 1 week.

---------

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>