NOISSUE - Add Invitations service layer (#124)
* feat(invitation): Add new function for sending invitations

This commit adds a new function for sending invitations in the invitation management and authorization code. The function allows users to send invitations to other users, providing necessary details such as email address and invitation message. This feature enhances the overall functionality of the application.

The code changes include the addition of the "sendInvitation" function in the invitation module, along with necessary validation checks and error handling. Additionally, test cases have been added to cover different scenarios and ensure the proper functioning of the feature.

This commit improves the invitation management and authorization functionality by enabling users to send invitations to others.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* test(invitations): Add test cases for invitation functions

Add new test cases to cover different scenarios and expected behavior in the invitation functions. These test cases ensure proper error handling and validate the functionality of the code related to invitations.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* feat: test with token

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* feat(invitations): update ListInvitations function

This commit updates the ListInvitations function in the invitations service. The changes include updating the page parameters to use InvitedByOrUserID instead of separate InvitedBy and UserID fields. This ensures that invitations are retrieved correctly based on the user ID.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* feat(invitations): Add rename to checkAdmin

An existing function has been renamed to "checkAdmin".

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* Refactor RetrieveAll method signature

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

---------

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>